MscmpSystSettings (mscmp_syst_settings v0.1.0)

A user options configuration management service.

The Settings Service provides caching and management functions for user configurable options which govern how the application operates. Multiple Settings Service instances may be in operation depending on the needs of the application; for example, in the case of multi-tenancy, each tenant will have its own instance of the Setting Service running since each tenant's needs of the application may unique.

On startup, the Settings Service creates an in memory cache and populates the cache from the database. Inquiries for settings are then served from the cache rather than the database as needed. Operations which change the Settings data are written to the database and then updated in the cache.

Settings maintained by this service may be changed by users at any time while the application is running. Therefore, any logic depending on the Settings from this service should be written as to be insensitive to such changes. Logic should avoid multiple retrievals of the same setting during any one transaction.

Runtime

child_spec(opts)

@spec child_spec(Keyword.t()) :: Supervisor.child_spec()

Returns a child specification for the Settings Service.

Parameters

opts - A keyword list of options. See the Options section for details.

Options

:debug (boolean/0) - If true, the GenServer backing the Settings Service will be started in debug mode.

:timeout (timeout/0) - Timeout value for the start_link call. The default value is :infinity.

:hibernate_after (timeout/0) - If present, the GenServer process awaits any message for the specified time before hibernating. The timeout value is expressed in Milliseconds.

:datastore_context_name (t:GenServer.name/0 ornil) - Specifies the name of the Datastore Context to be used by the Settings Service. *:service_name(t:GenServer.name/0 or nil) - The name to use for the GenServer backing this specific Settings Service instance.

Examples

iex> MscmpSystSettings.child_spec(
...>   service_name: MyApp.SettingsService,
...>   datastore_context_name: MyApp.DatastoreContext)
%{
  id: MscmpSystSettings.Runtime.Service,
  start:
    {MscmpSystSettings,
     :start_link,
     [MyApp.SettingsService, MyApp.DatastoreContext, [timeout: :infinity]]},
}

get_service()

@spec get_service() :: atom() | nil

Retrieve the current specific Settings Service name in effect for the process.

This function returns the name of the Settings Service that has been using the put_service/1 function to override the default Settings Service associated with the Instance Name. If no specific Settings Service name has been set, this function will return nil.

Examples

Retrieving a specific Settings Service name:

iex> MscmpSystSettings.put_service(:"MscmpSystSettings.TestSupportService")
...> MscmpSystSettings.get_service()
:"MscmpSystSettings.TestSupportService"

Retrieving a specific Settings Service name when no value is currently set for the process:

iex> MscmpSystSettings.put_service(nil)
...> MscmpSystSettings.get_service()
nil

put_service(settings_service_name)

@spec put_service(GenServer.name() | nil) :: GenServer.name() | nil

Establishes the currently active Settings Service instance for the current process.

Parameters

settings_service_name - the canonical name of the specific Settings Service to access. When this function is called with a non-nil argument, calls to Settings related functions will make use of the Settings Service specified here. Setting this value to nil will clear the currently active Settings Service for the process. Any calls to Settings related functions will fail until a new Settings Service is designated for the process using this function with a non-nil argument.

Examples

Setting a specific Settings Service name:

iex> MscmpSystSettings.put_service(:"MscmpSystSettings.TestSupportService")
...> MscmpSystSettings.get_service()
:"MscmpSystSettings.TestSupportService"

Clearing a previously set specific Service Name:

iex> MscmpSystSettings.put_service(nil)
...> MscmpSystSettings.get_service()
nil

refresh_from_database()

@spec refresh_from_database() :: :ok

Refreshes the cached settings values from the database.

Calling this function causes the existing cached settings to be purged from the cache and the cache to be repopulated from the database using the datastore context provided to start_link/1.

Examples

iex> MscmpSystSettings.refresh_from_database()
:ok

start_link(service_name, datastore_context_name, opts \\ [])

@spec start_link(
  MscmpSystSettings.Types.service_name(),
  MscmpSystDb.Types.context_service_name(),
  Keyword.t()
) :: {:ok, pid()} | {:error, MscmpSystError.t()}

Starts an instance of the Settings Service.

Starting the service establishes the required processes and pre-populates the service cache with data from the database. Most other functions in this module require that the service is started prior to use and will fail if the service is not started.

Parameters

service_name - The name of the Settings Service.

datastore_context_name - The name of the Datastore Context to be used by the Settings Service.

opts - Options for the Settings Service. See the Options section below for more information.

Options

:debug (boolean/0) - If true, the GenServer backing the Settings Service will be started in debug mode.

:timeout (timeout/0) - Timeout value for the start_link call. The default value is :infinity.

:hibernate_after (timeout/0) - If present, the GenServer process awaits any message for the specified time before hibernating. The timeout value is expressed in Milliseconds.

terminate_service()

@spec terminate_service() :: :ok

Terminates a running instance of the settings service.

Examples

> MscmpSystSettings.terminate_service()
:ok