MscmpSystSettings (mscmp_syst_settings v0.1.0)

MscmpSystSettings - User Configuration Management

A user options configuration management service.

The Settings Service provides caching and management functions for user configurable options which govern how the application operates. Multiple Settings Service instances may be in operation depending on the needs of the application; for example, in the case of multi-tenancy, each tenant will have its own instance of the Setting Service running since each tenant's needs of the application may unique.

On startup, the Settings Service creates an in memory cache and populates the cache from the database. Inquiries for settings are then served from the cache rather than the database as needed. Operations which change the Settings data are written to the database and then updated in the cache.

Settings maintained by this service may be changed by users at any time while the application is running. Therefore, any logic depending on the Settings from this service should be written as to be insensitive to such changes. Logic should avoid multiple retrievals of the same setting during any one transaction.

Runtime

get_settings_service()

@spec get_settings_service() :: MscmpSystSettings.Types.service_name() | nil

Retrieves the currently set Settings Service name or nil if none has been set.

Examples

iex> MscmpSystSettings.get_settings_service()
:settings_instance

put_settings_service(service_name)

@spec put_settings_service(MscmpSystSettings.Types.service_name() | nil) ::
  MscmpSystSettings.Types.service_name() | nil

Establishes the current Settings Service instance for the process.

A running system is likely to have more than one instance of the Settings Service running. For example, in multi-tenant applications each tenant may have its own instance of the Setting Service, caching data unique to the tenant.

Calling MscmpSystSettings.put_settings_service/1 will set the reference to the desired Settings Service instance for any subsequent MscmpSystSettings calls. The service name is set in the Process Dictionary of the process from which the calls are being made. As such, you must call put_settings_service at least once for any process from which you wish to access the Settings Service.

Because we're just thinly wrapping Process.put/2 here, the return value will be the previous value set here, or nil if no previous value was set.

Parameters

service_name - The name of the service which is to be set as servicing the process.

Examples

iex> MscmpSystSettings.put_settings_service(:settings_instance)

refresh_from_database()

@spec refresh_from_database() :: :ok

Refreshes the cached settings values from the database.

Calling this function causes the existing cached settings to be purged from the cache and the cache to be repopulated from the database using the datastore context provided to start_link/1.

Examples

iex> MscmpSystSettings.refresh_from_database()
:ok

start_link(params)

@spec start_link(
  {MscmpSystSettings.Types.service_name(), MscmpSystDb.Types.context_name()}
) ::
  {:ok, pid()} | {:error, MscmpSystError.t()}

Starts an instance of the Settings Service.

Starting the service establishes the required processes and pre-populates the service cache with data from the database. Most other functions in this module require that the service is started prior to use and will fail if the service is not started.

The service_name argument provides a unique name under which the service can be found. This argument is a subset of those that allowed for registering GenServers; the allowed forms for service name are simple atoms for basic local name registry or a "via tuple", such as might be used with the Registry module.

The datastore_context_name is an atom which represents a started MscmpSystDb context. This context will be used for accessing and modifying database data.

terminate_settings_service()

@spec terminate_settings_service() :: :ok

Terminates a running instance of the settings service.

Examples

> MscmpSystSettings.terminate_settings_service()
:ok